Prettier way but not all items rendered
