+++ src/components/App/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import styles from './App.css';
-import Form from '../Form/Form';
+import FormList from '../FormList/FormList';
 
 // eslint-disable-next-line react/prefer-stateless-function
 export default class App extends React.Component {
     render() {
         return (
             <div className={styles.mainPage}>
-                <Form />
+                <FormList />
             </div>
         );
     }
